api/requirements.txt
psycopg2-binary
gunicorn
python-decouple
django
djmail
cryptography
django-jinja
easy-thumbnails
raven
djangorestframework
markdown
django-filter
Unidecode
premailer
iso4217
pytz
pycountry
django-extensions
psd-tools
django-pglocks
html5lib
Markdown==3.1.1
diff-match-patch
social-auth-app-django
drf-link-header-pagination
django-redis 
channels 
channels-redis
django-otp 
qrcode[qrencode]
python-dateutil
celery
bleach==3.3.0
django-timezone-field
paramiko
django-cors-headers
geoip2
forex-python
jsonschema
